+++ src/contexts/UserContext.tsx
@@ -1,17 +1,27 @@
 import axios from 'axios'
 import { createContext, ReactNode, useContext, useState } from 'react'
 
+type trySignUpProps = {
+	navigate: (to: string) => void,
+	addr: string
+}
+
 export type UserContextData = {
-	username: string,
-    name: string,
-    password: string,
-    accessToken: string,
-    refreshToken: string,
-    setUsername: (s:string) => void,
-    setName: (s:string) => void,
-    setPassword: (s:string) => void,
-    setAccessToken: (s:string) => void,
-    setRefreshToken: (s:string) => void
+	name: string
+	setName: (s: string) => void
+	username: string
+	setUsername: (s: string) => void
+	password: string
+	setPassword: (s: string) => void
+	email: string
+	setEmail: (s: string) => void
+	birthday: Date
+	setBirthday: (d: Date) => void
+    isSaved: boolean,
+    setIsSaved: (b: boolean) =>void
+	isLoggedin: boolean
+	setIsLoggedin: (b: boolean) => void
+	trySignUp: (props: trySignUpProps) => void
 }
 
 export const UserContext = createContext<UserContextData | null>(null)
@@ -20,24 +30,49 @@ type ProviderProps = {
 	children: ReactNode
 }
 export function UserProvider({ children }: ProviderProps) {
+	const [name, setName] = useState('')
 	const [username, setUsername] = useState('')
-    const [name, setName] = useState('')
-    const [password, setPassword] = useState('')
-    const [accessToken, setAccessToken] = useState('')
-    const [refreshToken, setRefreshToken] = useState('')
+	const [password, setPassword] = useState('')
+	const [email, setEmail] = useState('')
+	const [birthday, setBirthday] = useState(new Date())
+    const [isSaved, setIsSaved] = useState(false)
+	const [isLoggedin, setIsLoggedin] = useState(false)
+	const trySignUp = async ({navigate, addr}: trySignUpProps) => {
+		try {
+			const response = await axios.post(
+				'/api/v1/auth/signup',
+				{
+					username: username,
+					name: name,
+					password: password,
+					contact: email,
+					contact_type: email
+				}
+			)
+			console.log(response)
+			navigate(addr)
+		} catch (error) {
+			alert('회원가입 실패')
+		}
+	}
 	return (
 		<UserContext.Provider
 			value={{
+				name,
+				setName,
 				username,
-                name,
-                password,
-                accessToken,
-                refreshToken,
-                setUsername,
-                setName,
-                setPassword,
-                setAccessToken,
-                setRefreshToken
+				setUsername,
+				password,
+				setPassword,
+				email,
+				setEmail,
+				birthday,
+				setBirthday,
+                isSaved,
+                setIsSaved,
+				isLoggedin,
+				setIsLoggedin,
+				trySignUp,
 			}}
 		>
 			{children}
