+++ src/pages/Login/Login.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 import { Link } from 'react-router-dom'
-import { useUserContext } from '../../contexts/AuthContext'
+import { useUserContext } from '../../contexts/UserContext'
 import { useEffect, useState } from 'react'
 import axios from 'axios'
