+++ src/pages/Login/signUp/Agree.tsx
@@ -1,6 +1,6 @@
 import { Link, useNavigate } from 'react-router-dom'
 import styled from 'styled-components'
-import { useUserContext } from '../../../contexts/AuthContext'
+import { useUserContext } from '../../../contexts/UserContext'
 
 const Img = styled.img`
 	width: 2rem;
